requirements.txt
streamlit
opencv-python
numpy
pillow
tensorflow
streamlit-webrtc